==> 1.Data_Cleaning.ipynb <==
# Fuel Economy
### Data on Cars used for Testing Fuel Economy
<hr>

The test data used to determine fuel economy estimates is derived from vehicle testing done at EPA's National Vehicle and Fuel Emissions Laboratory in Ann Arbor, Michigan, and by vehicle manufacturers who submit their own test data to EPA.

Each year, EPA provides fuel economy data to the Department of Energy (DOE), the Department of Transportation (DOT) and the Internal Revenue Service (IRS) so that they can administer their fuel economy-related programs.

In this scenario we will analyse how the fuel economy has made impact on vehicles after a decade. Therfor we will use the dataset for 2010 and 2020 respectively and analyse how much impact has been done. One of the major concern of recent times is pollution which contributes to environmental degradatation and vehicles adds the most of it. With the steep increase in vehicle buying we need to analyse the economy of vehicles, as the more fuel is burnt the more residue is emmisioned into the atmosphere plus the rapid decrease in non-renuable sources would put a devastation scenario to our future generations.


### About the Dataset
<hr>

Fuel economy data are the result of vehicle testing done at the Environmental Protection Agency's National Vehicle and Fuel Emissions Laboratory in Ann Arbor, Michigan, and by vehicle manufacturers with oversight by EPA.

#### Attribute Information

The data set comprises of the following features of vehicles and the respective vehicles data points :

* Model – vehicle make and model
* Displ – engine displacement in liters
* Cyl – number of engine cylinders
* Trans – transmission type plus number of gears
    1. Auto - Automatic
    2. Man - Manual
    3. SemiAuto - Semi-Automatic
    4. SCV - Selectable Continuously Variable (e.g. CVT with paddles)
    5. AutoMan - Automated Manual
    6. AMS - Automated Manual-Selectable (e.g. Automated Manual with paddles)
    7. Other - Other
    8. CVT - Continuously Variable
    9. CM3 - Creeper/Manual 3-Speed
    10. CM4 - Creeper/Manual 4-Speed
    11. C4 - Creeper/Manual 4-Speed
    12. C5 - Creeper/Manual 5-Speed
    13. Auto-S2 - Semi-Automatic 2-Speed
    14. Auto-S3 - Semi-Automatic 3-Speed
    15. Auto-S4 - Semi-Automatic 4-Speed
    16. Auto-S5 - Semi-Automatic 5-Speed
    17. Auto-S6 - Semi-Automatic 6-Speed
    18. Auto-S7 - Semi-Automatic 7-Speed
* Drive – 2-wheel Drive, 4-wheel drive/all-wheel drive
* Fuel – fuel(s)
* Veh Class – EPA vehicle class. 
* Air Pollution Score (Smog Rating) 
* City MPG – city fuel economy in miles per gallon
* Hwy MPG – highway fuel economy in miles per gallon
* Cmb MPG – combined city/highway fuel economy in miles per gallon
* Greenhouse Gas Score (Greenhouse Gas Rating) 
* SmartWay – Yes, No, or Elite. 
* Comb CO 2 – combined city/highway CO 2 tailpipe emissions in grams per mile





# import necessary libraries

import numpy as np
import pandas as pd

## Data Cleaning

# import data sets // both 2010 and 2020 economy data

data_2010 = pd.read_excel("data/all_alpha_10.xls")
data_2020 = pd.read_excel("data/all_alpha_20.xlsx")

# view data for 2010

data_2010.head()

# view data for 2020

data_2020.head()

# check shape od data set
print(data_2010.shape)
print(data_2020.shape)

# check columns

print(data_2010.columns)
print(data_2020.columns)

#### Few changes have to be made :

1. Data set for the year 2010 contains a column 'Sales Area'  and year 2020 contains the column 'Cert Region'. Which are same, after checking the values we should rename either one of them.


2. Year 2010 does not contain column 'Comb C02' whereas 2020 does therefor we will remove the column from 2020.

#### Check columns 'Sales Area' and 'Cert Region'.

data_2010['Sales Area'].value_counts()

data_2020['Cert Region'].value_counts()

### Drop Extraneous Columns

# data set 2010

data_2010.drop(['Stnd','Sales Area','Stnd Description','Underhood ID'], axis=1, inplace=True)

# data set 2010

data_2020.drop(['Stnd','Cert Region','Stnd Description','Underhood ID','Comb CO2'], axis=1, inplace=True)

# check weather all columns are same and of same structure

(data_2010.columns == data_2020.columns).all()

### Rename Columns

# replace space with _ and convert all letters to lowercase for ease
data_2010.rename(columns =lambda x: x.strip().lower().replace(' ','_'), inplace=True)

# replace space with _ and convert all letters to lowercase for ease
data_2020.rename(columns =lambda x: x.strip().lower().replace(' ','_'), inplace=True)

### Drop rows with missing values

# view missing value count for each feature in 2010
data_2010.isnull().sum()

# view missing value count for each feature in 2020
data_2020.isnull().sum()

# drop values
data_2010.dropna(inplace=True)
data_2020.dropna(inplace=True)

# check if any column in data 2010 and 2020 has null data
data_2010.isnull().sum().any()
data_2020.isnull().sum().any()

### Dedupe Data

# drop duplicates in both the datasets
data_2010.drop_duplicates(inplace=True)
data_2020.drop_duplicates(inplace=True)

# print number of duplicates again to confirm dedupe - should both be 0
print(data_2010.duplicated().sum())
print(data_2020.duplicated().sum())

### Fixing Data Types

# check data type for each columns
print(data_2010.info())
print(data_2020.info())

# check for data containing / in 2010
data_2010.fuel.str.contains('/').sum()

# check for data containing / in 2020
data_2020.fuel.str.contains('/').sum()

#### Observation

* We found that ___2010__ dataset contains __103__ and __2020__ dataset contains __134 hybrid vehicles__.

def change_dtype_intfloat(df,inputcolumns):
    '''

    Author : Niladri Ghosh
    Email : niladri1406@gmail.com


    The function takes in a single input, a list, which contains the columns names for the data set.
    Then the funtion makes two copies of dataframe since there are hybrid vehicles present in the 
    dataset, thus the data points contains a "/" to differentiate the values for each model. So we 
    copy two portions to different data frames by splitting it using "/". Finally combining both the
     data frames and convert each column to int respectively.


    '''
    # data frame for hybrid vehicles
    df_hb = df[df['fuel'].str.contains('/')]
    
    # create two copies
    df1 = df_hb.copy()
    df2 = df_hb.copy()
    
    inputcolumns = inputcolumns + ['fuel']
    
    for x in inputcolumns:
        
        df1[x] = df1[x].apply(lambda y: y.split("/")[0])
        df2[x] = df2[x].apply(lambda y: y.split("/")[1])
    
    # combine two data frames 
    new_df = df1.append(df2)
        
    # drop hybrid data drom original data frame
    df.drop(df_hb.index, inplace=True)
    
    # add new data frame to original data frame
    df = df.append(new_df, ignore_index=True)
    
    inputcolumns.remove('fuel')
    
    # change data type of columns
    for x in inputcolumns:
        try:
            df[x] = df[x].str.extract('(\d+)').astype(int)
        except:
            df[x] = df[x].str.extract('(\d+)').astype(float)
    
    return df

# convert column to int and float accordingly for 2010
ch_data_10 = ['air_pollution_score','city_mpg','hwy_mpg','cmb_mpg','greenhouse_gas_score']
df_2010 = change_dtype_intfloat(data_2010,ch_data_10)

df_2010.info()

df_2010[df_2010['air_pollution_score'].isna()==True]

#### Since few of the columns contained "mod / modified" unkown parameter data, we would remove them.

# remove values
df_2010.dropna(inplace=True)

# convert column to int and float accordingly for 2010
ch_data_20 = ['city_mpg','hwy_mpg','cmb_mpg']
df_2020 = change_dtype_intfloat(data_2020,ch_data_20)

# check data 
df_2020.info()

## Save cleaned data

### Save Data Sets
df_2010.to_csv('data/data_2010_v1.csv', index=False)
df_2020.to_csv('data/data_2020_v1.csv', index=False)
