==> 2.Feature_Engineering.ipynb <==
# Feature Engineering
<hr>

Feature engineering is the process of using domain knowledge to extract features from raw data via data mining techniques. These features can be used to improve the performance of machine learning algorithms. Feature engineering can be considered as applied machine learning itself.

Source : <a href="https://en.wikipedia.org/wiki/Feature_engineering#:~:text=Feature%20engineering%20is%20the%20process,as%20applied%20machine%20learning%20itself.\">Wikipedia</a>


### Process
<hr>
    
The feature engineering process is :

1. Brainstorming or testing features
2. Deciding what features to create
3. Creating features
4. Checking how the features work with your model
5. Improving your features if needed
6. Go back to brainstorming/creating more features until the work is done

# import necesary libraries

import numpy as np
import pandas as pd
from IPython.display import display
pd.options.display.max_columns = None

# import data set

df_2010 = pd.read_csv('data/data_2010_v1.csv')
df_2020 = pd.read_csv('data/data_2020_v1.csv')

def low_high(x):
    '''
    
    Author : Niladri Ghosh
    Email : niladri1406@gmail.com
    
    
    A function that takes in a single argument, a dataframe , then creates a extra column with 
    data "(column_name)_level" which describes whether the values in the data are low , medium, 
    mod_high or high. The values are calculated against the whole columns statistical data. We 
    use .describe() method to fetch the values for min, 25%, 50% and so on. For instance creating 
    a column for cmb_mpg (combined miles per gallon) values, the higher the value the more is the 
    mileage level.
    
    level          cmb_mpg_level   
    
    min - 25%      low        
    25% - 50%      medium    
    50% - 75%      mod_high      
    75% - max      high         
    
    
    '''
    
    
    for col in ['displ','cyl','air_pollution_score','city_mpg','hwy_mpg','cmb_mpg','greenhouse_gas_score']: 
        
        bin_edges = [x[col].describe()['min']-0.00001, x[col].describe()['25%']-0.00002, x[col].describe()['50%']-0.00001, 
                     x[col].describe()['75%'], x[col].describe()['max']]
        bin_names = ['low', 'medium', 'mod_high', 'high']
        x[col+"_level"] = pd.cut(x[col], bin_edges, labels=bin_names)

### Now we will create new rows of numerical data and divide it on the basis of scores if the values is between minimum value and 25% of the value -- it will be assigned low , similarly for values between 25% - 50% -- medium, 50% - 75% -- mod_high and 75% - Max -- high.

# data set 2010
low_high(df_2010)

# data set 2020
low_high(df_2020)

We need to perform __one hot encode__ on the newly created categorical data. By definition - "*__One Hot Encode refers to 
    splitting the column which contains numerical categorical data to many columns depending on the number of categories 
    present in that column. Each column contains “0” or “1” corresponding to which column it has been placed__*". For 
    example a column color has values red, blue and green (categorical data). Threfor the corresponding columns created 
    with values would be -
    
    color          color_red      color_blue    color_green
    
    red            1              0             0
    blue           0              1             0
    green          0              0             1
    blue           0              1             0
    red            1              0             0
    blue           0              1             0
    red            1              0             0
    green          0              0             1

# copy data set
data_2010_cp = df_2010.copy()

# rename columns
data_2010_cp.rename(columns = lambda  x : x + "_2010", inplace=True)

### Data 2010

# copy data set
df_2010_cp = data_2010_cp.copy()

# since model columns has unique data we'll be dropping it for the time being
df_2010_cp.drop('model_2010', axis=1, inplace=True)

# create onehot / dummies
df_2010_onehot = pd.get_dummies(df_2010_cp)

# insert columns model
df_2010_onehot.insert(0, 'model_2010' ,data_2010_cp['model_2010'])

df_2010_onehot.head(5)

### Dataset 2020

# create a copy of data
data_2020_cp = df_2020.copy()

# since model columns has unique data we'll be dropping it for the time being
data_2020_cp.drop('model', axis=1, inplace=True)

# create onehot / dummies
df_2020_onehot = pd.get_dummies(data_2020_cp)

# insert columns model
df_2020_onehot.insert(0, 'model' ,df_2020['model'])

df_2020_onehot.head(5)

## Combining Data set into one based on car models

Original data sets for 2010 and 2020, before combining we need to identify each data separately, therefor we will add a new column  __"year"__  which will determine the vehicle which year is it from. 

# merge data sets
df_combined = data_2010_cp.merge(df_2020, left_on="model_2010", right_on="model", how="inner")

### Combine onehot data set

df_combined_onehot = df_2010_onehot.merge(df_2020_onehot, left_on="model_2010", right_on="model", how="inner")

# check for null values
df_combined_onehot.isna().sum().any()

## Save Data set to csv files

# original data sets 2010 and 2020
df_2010.to_csv("data/data_2010_v2.csv", index=False)
df_2020.to_csv("data/data_2020_v2.csv", index=False)

# cobined data
df_combined.to_csv("data/data_combined.csv", index=False)

# combined onehot data
df_combined_onehot.to_csv("data/data_combined_onehot.csv", index=False)
