==> 4.Drawing_Conclusions_Combined.ipynb <==
# Results with Merged Dataset

### Q7: For all of the models that were produced in 2010 that are still being produced now, how much has the mpg improved and which vehicle improved the most?

# import necesary libraries

import numpy as np
import pandas as pd
from IPython.display import display

pd.options.display.max_columns = None

# load dataset
df_combined = pd.read_csv('data/data_combined.csv')

### 1. Create a new dataframe, `mpg_model`, that contain the mean combined mpg values in 2010 and 2020 for each unique model

# group by model, mean cmb_mpg_2010 and mean cmb_mpg for each.
mpg_model = df_combined.groupby('model').mean()[['cmb_mpg_2010', 'cmb_mpg']]

mpg_model.head()

### 2. Create a new column, `mpg_change`, with the change in mpg

# Subtract the mean mpg in 2010 from that in 2020 to get the change in mpg
mpg_model['mpg_change'] = model_mpg['cmb_mpg'] - model_mpg['cmb_mpg_2010']

mpg_model.head()

### 3. Find the vehicle that improved the most

# maximum change
max_change = mpg_model['mpg_change'].max()
max_change

# find details 
mpg_model[mpg_model['mpg_change'] == max_change]

### AUDI Q5 has a masive increase in mileage of 18.6 mpg.

### 4. Find the vehicle that improved the least

# maximum change
min_change = mpg_model['mpg_change'].min()
min_change

# find details 
mpg_model[mpg_model['mpg_change'] == min_change]

### Roush Stage 3 Mustang had a decrease in mileage of 3 mpg, and we certainly know why ! Its all muscle and power .

### 5. Find the vehicle that had no improvement

mpg_model[mpg_model['mpg_change']==0]

### There are 6 vehicle that had no change in thier mileage.
